Move psql code to own package
