+++ backend/main.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/Slowers-team/Slowers-App/application"
 	"github.com/Slowers-team/Slowers-App/database"
+	psqldatabase "github.com/Slowers-team/Slowers-App/database/psql"
 	"github.com/Slowers-team/Slowers-App/handlers"
 )
 
@@ -22,9 +23,9 @@ func main() {
 		}
 	}
 
-	var sqldb *database.SQLDatabase
+	var sqldb *psqldatabase.SQLDatabase
 	if useSQL == "true" {
-		sqldb := database.NewSQLDatabase(SQLDatabaseURI)
+		sqldb := psqldatabase.NewSQLDatabase(SQLDatabaseURI)
 		if env == "test" {
 			if err := sqldb.Connect("SlowersTest"); err != nil {
 				log.Fatal(err)
